[#8] feat: Cafe 생성 메소드 리팩토링
- Cafe.createCafe 메소드 리팩토링
- Cafe 기본적인 값 일치 테스트코드 추가
